+++ webPage/js/portal/src/platform.js
@@ -1,26 +1,22 @@
+const [workflowName,UpLoadList, UploadIndex] = [Symbol("workflowName"), Symbol("UpLoadList"), Symbol("UploadIndex")];
+window[UpLoadList] = [];
+
 $(function() {
     //弹窗出来后的一些操作
-    var input = document.getElementById("source");
-    //文件域选择文件时, 执行readFile函数,把上传的文档信息放置dataFile中
-    input.addEventListener('change', readFile, false);
-    var i = 1;
-    function readFile() {
-        var file = this.files[0];
-        console.log(file);
-        console.log("文件名:" + file.name);
-        console.log("文件类型:" + file.type);
-        console.log("文件大小:" + file.size);
-        var date = new Date();
-        var dateString = date.toLocaleDateString();
+    const input = document.getElementById("source");
 
-        //把每次上传的文档名字传入dataFile
-        var html = "<li><span>" + (i++) + "、</span><span>" + file.name + "</span><span>" + dateString + "</span><span class='closeFile'></span></li>";
-        $(".dataFile").append(html);
-        //点击关闭按钮，取消该项文档的添加
-        $(".closeFile").on("click", function() {
-            console.log($(this).parent("li"))
-            $(this).parent("li").remove();
+    //文件域选择文件时, 执行readFile函数,把上传的文档信息放置dataFile中
+    input.addEventListener("change", readFile, true);
+
+    function readFile() {
+        const file = this.files[0];
+
+        window[UpLoadList].push({
+          name : file.name,
+          file,
         });
+
+        $(".dataFile").html(buildUploadList());
     }
 
     // 点击反馈跳转到bbs
@@ -32,9 +28,15 @@ $(function() {
     $(".btn_close").on("click", function() {
         $(".popup").css("display", "none");
     });
-    
+
     //打开遮罩层
     $(".upload").on("click", function() {
+        const __self = $(".processActive").find("dt");
+        if (undefined === window[workflowName]) {
+          window[workflowName] = $(".aLink.Active").html();
+        }
+        const [nodeId, nodeName] = [__self.attr("data-id"), __self.find("span").html()];
+        $(".popup").find(".name").html(`${window[workflowName]} - ${nodeName} 更新`);
         $(".popup").css("display", "block");
     });
 
@@ -43,8 +45,9 @@ $(function() {
         console.log($(this).parent("div").parent("li"))
         $(this).parent("div").parent("li").remove();
     });
+
     //进入该页面，content_one从上往下运动
-    var width = $(window).width();
+    const width = $(window).width();
     if (width > 768) {
         var contentOne = $(".content_one").height();
         var height = -contentOne;
@@ -71,6 +74,93 @@ $(function() {
 
 });
 
+const uploadTask = (id, file, index) => new Promise((solve, reject) => {
+  const [xhr, formData] = [new XMLHttpRequest(), new FormData()];
+  formData.append("goods", file);
+  /*
+    xhr.timeout = 15 * 1000;
+    xhr.ontimeout = function (event) {
+        alert("请求超时");
+    };
+  */
+  xhr.open("POST", `/fs/update/${id}`);
+  xhr.setRequestHeader("Accept", "application/json");
+  xhr.send(formData);
+  xhr.onreadystatechange = () => {
+      if (4 === xhr.readyState) {
+          if (200 === xhr.status) {
+              try {
+                const result = JSON.parse(xhr.responseText);
+
+                if (200 !== result.code) {
+                  reject({
+                    msg: result.error,
+                    code: result.code,
+                  });
+                  return ;
+                }
+
+                Object.assign(result, {
+                  [UploadIndex]: index,
+                });
+
+                solve(result);
+
+              } catch(err) {
+                reject(xhr.responseText);
+              }
+          }
+          reject();
+      }
+  };
+});
+
+const uploadTasks = nodeId => {
+    const tasks = window[UpLoadList].map((object, index) => uploadTask(nodeId, object.file, index));
+    return Promise.all(tasks);
+};
+
+const beginUpload = () => {
+    const nodeId = $(".processActive").find("dt").attr("data-id");
+    uploadTasks(nodeId).then(data => {
+        window[UpLoadList].length = 0;
+        $(".popup").css("display", "none");
+        $(".dataFile").html("");
+        $.ajax({
+          type: "GET",
+          url: `/workflow/${nodeId}/files`,
+          dataType: "json",
+          beforeSend: xhr => xhr.setRequestHeader("accept","application/json"),
+          success: result => {
+            if (401.5 === result.code) {
+              alert("登录许可已失效，请重新获取登录许可");
+              location.href = "/portal/login";
+              return ;
+            }
+            if (200 !== result.code) {
+              alert("服务器繁忙");
+            } else {
+              buildFiles(result.result.produceList);
+            }
+          }
+        });
+    }).catch(err => {
+        err.code === 403? alert("非该项目组成员禁止上传文件") : alert("上传失败 " + err.msg);
+    });
+};
+
+const delUploadFileList = btn => {
+    window[UpLoadList].splice($(btn).attr("data-index"), 1);
+    $(".dataFile").html(buildUploadList());
+};
+
+const buildUploadList = () => window[UpLoadList].map((file, index) => `<li>
+  <span>${index + 1}</span>
+  <span>${file.name}</span>
+  <span>${new Date().toLocaleDateString()}</span>
+  <span class="closeFile" data-index=${index} onClick="javascript:delUploadFileList(this);"><span>
+</li>`);
+
 const downLoadFile = (btn) => {
   window.open("/fs/download/" + $(btn).attr("data-id"));
 };
@@ -81,11 +171,22 @@ const cleanSelect = () => {
   });
 };
 
-const cleanSelectd = () => {
+const cleanSelectd = (nodeName) => {
   $("dl").attr("class", "");
+  $(".data_title").html(`${nodeName} 资料`);
 }
 
+const buildUploadItem = (allowedUpload = false) => {
+  const uploadButton = $("#uploadButton");
+  if (allowedUpload) {
+    uploadButton.hasClass("upload") ? "": uploadButton.addClass("upload");
+  } else {
+    uploadButton.hasClass("upload") ? uploadButton.removeClass("upload"): "";
+  }
+};
+
 const buildProcess = (nodeList, actionNum) => {
+  $(".data_title").html(`${nodeList[actionNum].name} 资料`);
   const html = nodeList.map((node, index) => `<dl class=${actionNum === index? "processActive":""}>
     <dt data-id=${node._id} onClick="javascript:showFiles(this);">
       <a class="tabIcon tabIcon_plan" href="javascript:void(0);"></a>
@@ -123,16 +224,22 @@ const buildFiles = (files) => {
 
 const showFiles = btn => {
   const id = $(btn).attr("data-id");
+  const nodeName = $(btn).find("span").html();
   $.ajax({
     type: "GET",
     url: `/workflow/${id}/files`,
     dataType: "json",
     beforeSend: xhr => xhr.setRequestHeader("accept","application/json"),
     success: result => {
+      if (401.5 === result.code) {
+        alert("登录许可已失效，请重新获取登录许可");
+        location.href = "/portal/login";
+        return ;
+      }
       if (200 !== result.code) {
         alert("服务器繁忙");
       } else {
-        cleanSelectd();
+        cleanSelectd(nodeName);
         $(btn).parent().attr("class", "processActive");
         buildFiles(result.result.produceList);
       }
@@ -142,12 +249,18 @@ const showFiles = btn => {
 
 const selectWorkflow = btn => {
   const _id = $(btn).attr("data-id");
+  window[workflowName] = $(btn).find("span").html();
   $.ajax({
     type: "GET",
     url: `/workflow/${_id}/simple`,
     dataType: "json",
     beforeSend: xhr => xhr.setRequestHeader("accept","application/json"),
     success: result => {
+      if (401.5 === result.code) {
+        alert("登录许可已失效，请重新获取登录许可");
+        location.href = "/portal/login";
+        return ;
+      }
       if (200 !== result.code) {
         alert("服务器繁忙");
       } else {
@@ -155,6 +268,7 @@ const selectWorkflow = btn => {
         $(btn).find("span").css({color: "#3d9bff"});
         buildProcess(result.result.nodeList, result.result.status.index);
         buildFiles(result.result.status.produceList);
+        buildUploadItem(result.result.allowedUpload);
       }
     }
   });